Major refactoring - Splitting responsibility in data source services, mapping services and   data services - Reorganize package structur to reflect data sources, mapping and data - Tried to use same name on same stuff
Added config for endpoint address url.
Updated README.
