+++ src/main/java/no/fint/p360/data/exception/CreateDocumentException.java
@@ -1,0 +1,4 @@
+package no.fint.p360.data.exception;
+
+public class CreateDocumentException extends RuntimeException {
+}
