+++ src/main/java/no/fint/p360/data/exception/GetDocumentException.java
@@ -1,0 +1,7 @@
+package no.fint.p360.data.exception;
+
+public class GetDocumentException extends RuntimeException {
+    public GetDocumentException(String message) {
+        super(message);
+    }
+}
