+++ src/main/java/no/fint/p360/data/noark/journalstatus/JournalStatusService.java
@@ -1,0 +1,29 @@
+package no.fint.p360.data.noark.journalstatus;
+
+import lombok.extern.slf4j.Slf4j;
+import no.fint.model.resource.administrasjon.arkiv.JournalStatusResource;
+import no.fint.p360.data.p360.P360SupportService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+import java.util.stream.Stream;
+
+@Slf4j
+@Service
+public class JournalStatusService {
+
+    @Autowired
+    private P360SupportService supportService;
+
+    @Autowired
+    private JournalStatusFactory factory;
+
+    @Value("${fint.p360.tables.journal-status:Journal status}")
+    private String journalStatusTable;
+
+    public Stream<JournalStatusResource> getJournalStatusTable() {
+        return supportService.getCodeTableRowResultStream(journalStatusTable)
+                .map(factory::toFintResource);
+    }
+}
