+++ src/main/java/no/fint/p360/data/noark/part/PartFactory.java
@@ -1,4 +1,4 @@
-package no.fint.p360.data.fint;
+package no.fint.p360.data.noark.part;
 
 import no.fint.arkiv.p360.contact.ContactPersonResult;
 import no.fint.arkiv.p360.contact.EnterpriseResult;
