+++ src/main/java/no/fint/p360/data/utilities/RequestUtilities.java
@@ -17,4 +17,9 @@ public class RequestUtilities {
         map.put(BindingProvider.USERNAME_PROPERTY, appProps.getP360User());
         map.put(BindingProvider.PASSWORD_PROPERTY, appProps.getP360Password());
     }
+
+    public void setEndpointAddress(Map<String, Object> map, String service) {
+        map.put(BindingProvider.ENDPOINT_ADDRESS_PROPERTY,
+                String.format("%s/SI.WS.Core/SIF/%s.svc", appProps.getEndpointBaseUrl(), service));
+    }
 }
